requirements.txt
av==12.3.0
certifi==2024.7.4
charset-normalizer==3.3.2
colorama==0.4.6
coloredlogs==15.0.1
ctranslate2==4.3.1
decorator==4.4.2
faster-whisper==1.0.3
ffmpeg-python==0.2.0
filelock==3.15.4
flatbuffers==24.3.25
fsspec==2024.6.1
future==1.0.0
huggingface-hub==0.24.5
humanfriendly==10.0
idna==3.7
imageio==2.34.2
imageio-ffmpeg==0.5.1
Jinja2==3.1.4
llvmlite==0.43.0
MarkupSafe==2.1.5
more-itertools==10.4.0
moviepy==1.0.3
mpmath==1.3.0
networkx==3.3
numba==0.60.0
numpy==1.26.4
onnxruntime==1.18.1
openai-whisper==20231117
packaging==24.1
pillow==10.4.0
proglog==0.1.10
protobuf==5.27.3
pyreadline3==3.4.1
PyYAML==6.0.2
regex==2024.7.24
requests==2.32.3
sympy==1.13.1
tiktoken==0.7.0
tokenizers==0.20.0
torch==2.4.0
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==2.2.2
